requirements.txt
langchain-cerebras
twilio
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
redis[hiredis]
pydantic-redis